+++ packages/elements/src/components/ui/buttons/menu-button/MenuButton.module.css
@@ -1,0 +1,92 @@
+.menuButton {
+  --swui-menu-button-item-height: 40px;
+  --swui-menu-button-separator-color: var(--swui-white);
+  --swui-menu-button-text-color: var(--lhds-color-ui-900);
+  --swui-menu-button-icon-color: var(--lhds-color-ui-900);
+  --swui-menu-button-background-color: transparent;
+  --swui-menu-button-background-color-hover: var(--lhds-color-ui-400);
+  --swui-menu-button-background-color-focus: var(--lhds-color-ui-400);
+  --swui-menu-button-background-color-active: var(--lhds-color-ui-300);
+  --swui-menu-button-background-selected: var(--lhds-color-blue-100);
+  --swui-menu-button-background-selected-hover: var(--lhds-color-blue-200);
+  --swui-menu-button-background-selected-focus: var(--lhds-color-blue-200);
+  --swui-menu-button-background-selected-active: var(--lhds-color-blue-100);
+
+  /* State vars */
+  --current-separator-color: var(--swui-menu-button-separator-color);
+  --current-text-color: var(--swui-menu-button-text-color);
+  --current-icon-color: var(--swui-menu-button-icon-color);
+  --current-background-color: var(--swui-menu-button-background-color);
+  --current-background-color-hover: var(
+    --swui-menu-button-background-color-hover
+  );
+  --current-background-color-focus: var(
+    --swui-menu-button-background-color-focus
+  );
+  --current-background-color-active: var(
+    --swui-menu-button-background-color-active
+  );
+  --current-background-selected: var(--swui-menu-button-background-selected);
+  --current-background-selected-hover: var(
+    --swui-menu-button-background-selected-hover
+  );
+  --current-background-selected-focus: var(
+    --swui-menu-button-background-selected-focus
+  );
+  --current-background-selected-active: var(
+    --swui-menu-button-background-selected-active
+  );
+
+  /* Styling */
+
+  background-color: var(--current-background-color);
+
+  .label {
+    font-size: var(--current-text-size, var(--swui-font-size-medium));
+  }
+
+  button {
+    --swui-button-text-color: var(--swui-white);
+    --swui-button-background-color-hover: var(
+      --swui-menu-button-background-color-hover
+    );
+
+    border: 0;
+    padding: 0;
+    background: var(--current-background-color);
+    height: var(--swui-menu-button-item-height);
+
+    [data-collapsed] [aria-expanded="true"] &,
+    &:hover {
+      background: var(--current-background-color-hover);
+    }
+
+    &:focus {
+      outline: none;
+    }
+
+    &:focus-visible {
+      background: var(--current-background-color-active);
+    }
+
+    &:active {
+      background: var(--current-background-color-active);
+    }
+
+    &.selected {
+      background: var(--current-background-selected);
+
+      &:hover {
+        background: var(--current-background-selected-hover);
+      }
+
+      &:focus-visible {
+        background: var(--current-background-selected-focus);
+      }
+
+      &:active {
+        background: var(--current-background-selected-active);
+      }
+    }
+  }
+}
