+++ packages/elements/src/index.ts
@@ -17,6 +17,7 @@ export * from "./components/ui/buttons/PrimaryButton";
 export * from "./components/ui/buttons/SecondaryButton";
 export * from "./components/ui/buttons/common/ButtonCommon";
 export * from "./components/ui/buttons/common/ButtonContent";
+export * from "./components/ui/buttons/menu-button/MenuButton";
 export * from "./components/ui/button-links/FlatButtonLink";
 export * from "./components/ui/button-links/PrimaryButtonLink";
 export * from "./components/ui/button-links/SecondaryButtonLink";
